requirements.txt
transformers
sentence-transformers
faiss-cpu
bs4
requests
pdfplumber
python-docx
